Move MQTT and InfluxDB to their own header files
