+++ src/display.h
@@ -179,7 +179,7 @@ void Displaymachinestate() {
                 u8g2.print(wifiReconnects);
             }
 
-            if (MQTT == 1) {
+            #if MQTT
                 if (mqtt.connected() == 1) {
                     u8g2.setCursor(77, 1);
                     u8g2.setFont(u8g2_font_profont11_tf);
@@ -188,7 +188,7 @@ void Displaymachinestate() {
                     u8g2.setCursor(77, 2);
                     u8g2.print("");
                 }
-            }
+            #endif
         } else {
             u8g2.setCursor(40, 2);
             u8g2.print(langstring_offlinemod);
